requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==3.0.6
